--- src/types/layout.ts
@@ -1,6 +1,0 @@
-export interface ProfileMenuItem {
-  id: string
-  label: string
-  onSelect: () => void
-  tone?: 'default' | 'danger'
-}
